Fix number_of_selfloops() call for recent networkx versions
The number_of_selfloops() graph method was deprecated in the networkx
2.0 release[1] which was released in 2017 and replaced with a
function that provides the same functionality. In the recent 2.4 release
this deprecated method has been removed. This commit updates the use of
number_of_selfloops to use that function instead of the deprecated and
now removed method. It also increases the minimum version supported by
the library to use 2.0 which was the version this function was
introduced.

Fixes #14

[1] https://networkx.github.io/documentation/stable/release/release_2.0.html#api-changes